Implement OpenMP
Implemented a simple OpenMP parallel for loop over the main loop.
There are now two launch arguments. The first one is the path to the
directory to store the save file and the second (optional) argument
indicate the number of OpenMP threads to use. If not specified it
defaults to the maximum number allowed by OpenMP

- Updated the makefile for OpenMP and setting the compiler path
  semi-automatically
- Updated the readme with new launch instructions
- Changed how the savefile took paths to work with OpenMP
- Add a more sophisticated class for timing